requirements.txt
streamlit==1.30.0
pandas==2.0.3
openpyxl==3.1.2
plotly==5.17.0
streamlit-aggrid==0.3.4.post3
streamlit-option-menu==0.3.6
bcrypt==4.0.1
streamlit-extras==0.3.5
extra-streamlit-components==0.1.60
streamlit-shadcn-ui==0.1.2
streamlit-lottie==0.0.5
reportlab==4.0.4
Pillow>=10.2.0
email-validator==2.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Librerías para Supabase y PostgreSQL
psycopg2-binary==2.9.10
supabase==2.18.1
python-dotenv==1.1.1

# Librerías adicionales útiles para Supabase
schedule==1.2.0